Updating Environmental Methods
