+++ Revit_Engine/Modify/Messages/Warnings.cs
@@ -1,0 +1,24 @@
+﻿using Autodesk.Revit.DB;
+using Autodesk.Revit.DB.Structure;
+using BH.oM.Base;
+using System.Collections.Generic;
+using System.Reflection;
+using System;
+
+namespace BH.Engine.Revit
+{
+
+    /// <summary>
+    /// BHoM Revit Engine Modify Methods
+    /// </summary>
+    public static partial class Modify
+    {
+        /***************************************************/
+        /****              Private methods              ****/
+        /***************************************************/
+
+
+
+        /***************************************************/
+    }
+}
