+++ Revit_Engine/Revit_Engine.csproj
@@ -103,7 +103,8 @@
     <Compile Include="Convert\Environment\ToRevit\CompoundStructureLayer.cs" />
     <Compile Include="Convert\Environment\ToRevit\Element.cs" />
     <Compile Include="Convert\Environment\ToRevit\ElementType.cs" />
-    <Compile Include="Convert\Errors.cs" />
+    <Compile Include="Convert\Messages\Warnings.cs" />
+    <Compile Include="Convert\Messages\Errors.cs" />
     <Compile Include="Convert\Geometry\ToBHoM\Curve.cs" />
     <Compile Include="Convert\Geometry\ToBHoM\Edge.cs" />
     <Compile Include="Convert\Geometry\ToBHoM\LocationPoint.cs" />
@@ -129,7 +130,11 @@
     <Compile Include="Create\RevitSettings.cs" />
     <Compile Include="Create\SelectionSettings.cs" />
     <Compile Include="Create\WorksetSettings.cs" />
+    <Compile Include="Modify\AddBHoMObject.cs" />
+    <Compile Include="Modify\AddBHoMObjects.cs" />
     <Compile Include="Modify\AssignAdjacentSpaces.cs" />
+    <Compile Include="Modify\Messages\Errors.cs" />
+    <Compile Include="Modify\Messages\Warnings.cs" />
     <Compile Include="Modify\RemoveIdentifiers.cs" />
     <Compile Include="Modify\SetCustomData.cs" />
     <Compile Include="Modify\SetIdentifiers.cs" />
@@ -146,8 +151,11 @@
     <Compile Include="Query\Element.cs" />
     <Compile Include="Query\ElementId.cs" />
     <Compile Include="Query\ElementIds.cs" />
+    <Compile Include="Query\GetEnergyAnalysisModel.cs" />
     <Compile Include="Query\IntegerValue.cs" />
     <Compile Include="Query\MaterialGrade.cs" />
+    <Compile Include="Query\Messages\Errors.cs" />
+    <Compile Include="Query\Messages\Warnings.cs" />
     <Compile Include="Query\Outline.cs" />
     <Compile Include="Query\SpatialElementId.cs" />
     <Compile Include="Query\SpatialElementIds.cs" />
